+++ .gitignore
@@ -7,4 +7,4 @@ build/
 out/
 
 # Ignore environment files
-.env
+# .env
