+++ src/server.ts
@@ -1,7 +1,7 @@
 import http from 'http';
 import url from 'url';
-import { v4 as uuidv4, validate as uuidValidate } from 'uuid';
 import { User } from './types/types';
+import { getAllUsers, getUserById, createNewUser, updateUser, deleteUser } from './user/userMethods';
 
 let users: User[] = [];
 
@@ -9,123 +9,17 @@ export const server = http.createServer((req, res) => {
   const { pathname } = url.parse(req.url || '', true);
 
   if (req.method === 'GET' && pathname === '/api/users') {
-    // Get all users
-    res.writeHead(200, { 'Content-Type': 'application/json' });
-    res.end(JSON.stringify(users));
+    getAllUsers(res, users);
   } else if (req.method === 'GET' && pathname?.startsWith('/api/users/')) {
-    // Get user by ID
-
-    const userId = pathname.split('/').pop()!;
-
-    if (!uuidValidate(userId)) {
-      res.writeHead(400, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify({ error: 'Invalid userId' }));
-      return;
-    }
-
-    const user = users.find((user) => user.id === userId);
-
-    if (!user) {
-      res.writeHead(404, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify({ error: 'User not found' }));
-      return;
-    }
-
-    res.writeHead(200, { 'Content-Type': 'application/json' });
-    res.end(JSON.stringify(user));
+    getUserById(res, users, pathname);
   } else if (req.method === 'POST' && pathname === '/api/users') {
-    // Create a new user
-    let body = '';
-
-    req.on('data', (chunk) => {
-      body += chunk.toString();
-    });
-
-    req.on('end', () => {
-      let parsedData: User;
-
-      try {
-        parsedData = JSON.parse(body);
-      } catch (error) {
-        res.writeHead(400, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify({ error: 'Invalid data in request' }));
-        return;
-      }
-      const { name, age, hobbies } = parsedData;
-
-      if (!name || !age || !hobbies) {
-        res.writeHead(400, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify({ error: 'Missing required fields' }));
-        return;
-      }
-
-      const newUser: User = {
-        id: uuidv4(),
-        name,
-        age,
-        hobbies,
-      };
-
-      users.push(newUser);
-
-      res.writeHead(201, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify(newUser));
-    });
+    createNewUser(res, req, users);
   } else if (req.method === 'PUT' && pathname?.startsWith('/api/users/')) {
-
-    const userId = pathname.split('/').pop()!;
-
-    if (!uuidValidate(userId)) {
-      res.writeHead(400, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify({ error: 'Invalid userId' }));
-      return;
-    }
-
-    let body = '';
-
-    req.on('data', (chunk) => {
-      body += chunk.toString();
-    });
-
-    req.on('end', () => {
-      const userIndex = users.findIndex((user) => user.id === userId);
-
-      if (userIndex === -1) {
-        res.writeHead(404, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify({ error: 'User not found' }));
-        return;
-      }
-
-      const updatedUser = { ...users[userIndex], ...JSON.parse(body) };
-      users[userIndex] = updatedUser;
-
-      res.writeHead(200, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify(updatedUser));
-    });
+    updateUser(res, req, users, pathname);
   } else if (req.method === 'DELETE' && pathname?.startsWith('/api/users/')) {
-    // Delete user by ID
-    const userId = pathname.split('/').pop()!;
-
-    if (!uuidValidate(userId)) {
-      res.writeHead(400, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify({ error: 'Invalid userId' }));
-      return;
-    }
-
-    const userIndex = users.findIndex((user) => user.id === userId);
-
-    if (userIndex === -1) {
-      res.writeHead(404, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify({ error: 'User not found' }));
-      return;
-    }
-
-    users.splice(userIndex, 1);
-
-    res.writeHead(204, { 'Content-Type': 'application/json' });
-    res.end();
+    deleteUser(res, users, pathname);
   } else {
-    // Handle 404
+    // Handle errors
     res.writeHead(404, { 'Content-Type': 'application/json' });
     res.end(JSON.stringify({ error: 'Invalid data in request' }));
   }
